+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+- Change sigma_proposals-input in metropolis from list to dict
 - Fix acq_noise_var-bug in acquisition.py. Influenced BOLFI.
 
 0.8.3 (2021-02-17)